refactor: 'CreditCardColorPicker' component
- 'CreditCardCompanyPicker'로 이름 변경
해당 컴포넌트의 기능과 이름이 의미적으로 일치하도록 변경
- onClick으로 prop 이름 변경
카드 회사를 선택하는 일반적인 동작을 수행하는 버튼이므로 onClick으로 변경
