+++ src/components/organisms/index.tsx
@@ -1,3 +1,3 @@
 export { default as CreditCard } from '@/components/organisms/CreditCard';
-export { default as CreditCardColorPicker } from '@/components/organisms/CreditCardColorPicker';
+export { default as CreditCardCompanyPicker } from '@/components/organisms/CreditCardCompanyPicker';
 export { default as CreditCardTextFields } from '@/components/organisms/CreditCardTextFields';
